+++ tools/launch.sh
@@ -14,7 +14,7 @@ function stop_service {
     cd "${apps_dir}/${service}"
     profile="$(get_MP_D_PROFILE_x "${config_file}" "${service}")"
     # shellcheck disable=SC2086
-    docker-compose --project-name "${profile}" stop
+    docker compose --project-name "${profile}" stop
   )
 }
 
@@ -31,7 +31,7 @@ function run_mariadb {
       cd "${apps_dir}/mariadb"
       profile="$(get_MP_D_PROFILE_x "${config_file}" "mariadb")"
       # shellcheck disable=SC2086
-      docker-compose --project-name "${profile}" ${mp_docker_command}
+      docker compose --project-name "${profile}" ${mp_docker_command}
     )
   fi
 }
@@ -51,7 +51,7 @@ function run_antivirus {
       cd "${apps_dir}/${service}"
       profile="$(get_MP_D_PROFILE_x "${config_file}" "${service}")"
       # shellcheck disable=SC2086
-      docker-compose --project-name "${profile}" ${mp_docker_command} --build
+      docker compose --project-name "${profile}" ${mp_docker_command} --build
     )
   fi
 }
@@ -71,7 +71,7 @@ function run_fail2ban {
       cd "${apps_dir}/${service}"
       profile="$(get_MP_D_PROFILE_x "${config_file}" "${service}")"
       # shellcheck disable=SC2086
-      docker-compose --project-name "${profile}" ${mp_docker_command} --build
+      docker compose --project-name "${profile}" ${mp_docker_command} --build
     )
   fi
 }
@@ -88,10 +88,10 @@ function run_mail {
     profile="$(get_MP_D_PROFILE_x "${config_file}" "mail")"
     if [[ "${mp_docker_command}" == "restart" ]]; then
       # shellcheck disable=SC2086
-      docker-compose --project-name "${profile}" ${mp_docker_command}
+      docker compose --project-name "${profile}" ${mp_docker_command} --force-recreate
     else
       # shellcheck disable=SC2086
-      docker-compose --project-name "${profile}" ${mp_docker_command} --build
+      docker compose --project-name "${profile}" ${mp_docker_command} --build --force-recreate
     fi
   )
 }
@@ -111,7 +111,7 @@ function run_web {
       cd "${apps_dir}/web"
       profile="$(get_MP_D_PROFILE_x "${config_file}" "web")"
       # shellcheck disable=SC2086
-      docker-compose --project-name "${profile}" ${mp_docker_command} mail
+      docker compose --project-name "${profile}" ${mp_docker_command} mail --force-recreate
     )
   fi
 
@@ -124,10 +124,10 @@ function run_web {
       profile="$(get_MP_D_PROFILE_x "${config_file}" "web")"
       if [[ "${mp_docker_command}" == "restart" ]]; then
         # shellcheck disable=SC2086
-        docker-compose --project-name "${profile}" ${mp_docker_command} pa
+        docker compose --project-name "${profile}" ${mp_docker_command} pa --force-recreate
       else
         # shellcheck disable=SC2086
-        docker-compose --project-name "${profile}" ${mp_docker_command} --build pa
+        docker compose --project-name "${profile}" ${mp_docker_command} --build pa --force-recreate
       fi
     )
   fi
@@ -140,7 +140,7 @@ function run_web {
       cd "${apps_dir}/web"
       profile="$(get_MP_D_PROFILE_x "${config_file}" "web")"
       # shellcheck disable=SC2086
-      docker-compose --project-name "${profile}" ${mp_docker_command} pma
+      docker compose --project-name "${profile}" ${mp_docker_command} pma --force-recreate
     )
   fi
 
@@ -150,7 +150,7 @@ function run_web {
       cd "${apps_dir}/web"
       profile="$(get_MP_D_PROFILE_x "${config_file}" "web")"
       # shellcheck disable=SC2086
-      docker-compose --project-name "${profile}" ${mp_docker_command} nginx
+      docker compose --project-name "${profile}" ${mp_docker_command} nginx --force-recreate
     )
   fi
 }